move time-info validationStatus to root
