--- src/components/ShareExperience/TimeSalaryForm/formElements.js
@@ -1,4 +1,0 @@
-export const DAY_PROMISED_WORK_TIME = 'dayPromisedWorkTime';
-export const DAY_REAL_WORK_TIME = 'dayRealWorkTime';
-export const WEEK_WORK_TIME = 'weekWorkTime';
-export const OVERTIME_FREQUENCY = 'overtimeFrequency';
